+++ src/components/Callout/Callout.tsx
@@ -3,11 +3,10 @@ import { MdClose } from 'react-icons/md';
 import { bem } from '../../utils';
 import { Context } from '../../constants';
 import styles from './Callout.scss';
-import { NotEmptyReactNode } from '../../customTypes/types';
 
 export interface Props extends React.HTMLAttributes<HTMLDivElement> {
     /** Content to be rendered inside the container */
-    children: NotEmptyReactNode;
+    children: React.ReactNode;
     /** The Callout context (e.g. info, danger, success etc. - defaults to info) */
     context?: Context;
     /** A function to be called when close button was clicked */
@@ -22,6 +21,10 @@ export const Callout: React.FC<Props> = ({
     children,
     ...rest
 }) => {
+    if (!children) {
+        return null;
+    }
+
     return (
         <div {...rest} {...block({ context, ...rest })}>
             <div {...elem('content', { context })}>{children}</div>