+++ src/components/Icon/IconHarvester/__tests__/IconHarvester.spec.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import { IconHarvester } from '../IconHarvester';
+
+describe('<IconHarvester>', () => {
+    it('should render an Harvester icon', () => {
+        const view = render(<IconHarvester />);
+        expect(view.container).toMatchSnapshot();
+    });
+});