+++ src/components/Icon/IconSourcebox/__tests__/IconSourcebox.spec.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import { IconSourcebox } from '../IconSourcebox';
+
+describe('<IconSourcebox>', () => {
+    it('should render an Sourcebox icon', () => {
+        const view = render(<IconSourcebox />);
+        expect(view.container).toMatchSnapshot();
+    });
+});