+++ src/components/Icon/LogoTextkernel/__tests__/IconTextkernelFull.spec.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import { LogoTextkernel } from '../LogoTextkernel';
+
+describe('<LogoTextkernel>', () => {
+    it('should render an Textkernel icon', () => {
+        const view = render(<LogoTextkernel />);
+        expect(view.container).toMatchSnapshot();
+    });
+});