+++ src/components/Text/Text.tsx
@@ -2,11 +2,10 @@ import * as React from 'react';
 import { bem } from '../../utils';
 import styles from './Text.scss';
 import { Size, Context } from '../../constants';
-import { NotEmptyReactNode } from '../../customTypes/types';
 
 export interface Props extends React.HTMLAttributes<HTMLElement> {
     /** Text content */
-    children: NotEmptyReactNode;
+    children: React.ReactNode;
     /** Text should be rendered inline */
     inline?: boolean;
     /** The context of the text, effecting its color (e.g. primary, info, danger, success etc. 'neutral' added as special context here) */
@@ -19,6 +18,10 @@ const { block } = bem('Text', styles);
 
 export const Text = React.forwardRef<HTMLElement, Props>(
     ({ children, context = 'default', inline = false, size = 'normal', ...rest }, ref) => {
+        if (!children) {
+            return null;
+        }
+
         const HtmlNodeType = inline ? 'span' : 'p';
 
         return (